Adição de nova posiblidade ao adicionar/editar dados de veículo, correção de bugs e melhoria no readme.md

+++ routes/vehiclesRoutes.js
@@ -10,54 +10,85 @@ router.get("/vehicle/:id", async (req, res) => {
    await Vehicles.listVehicle(req, res);
 });
 
-router.get("/vehicles/:status?/:page?", async (req, res) => {
+router.get("/vehicles/:status/:page?", async (req, res) => {
    await Vehicles.listVehicles(req, res);
 });
 
-router.get("/vehiclesByType/:type?/:page?", async (req, res) => {
+router.get("/vehiclesByType/:type/:page?", async (req, res) => {
    await Vehicles.listVehiclesByType(req, res);
 });
 
-router.get("/vehiclesByModel/:model?/:page?", async (req, res) => {
+router.get("/vehiclesByModel/:model/:page?", async (req, res) => {
    await Vehicles.listVehiclesByModel(req, res);
 });
 
 router.post("/vehicles", login, upload.single("img"), validationRules, async (req, res) => {
 
-   const { path: imgPath } = req.file;
+   // Verifica se foi passada uma img via form
+   if(!req.file){
 
-   // Guarda os erros de validação
-   const validationErros = validationResult(req);
+      // Guarda os erros de validação
+      const validationErros = validationResult(req);
+         
+      // Verifica se ocorreu algum erro
+      if(!validationErros.isEmpty()){
+         return res.status(400).json({ errors: validationErros.array() });
+      }
 
-   // Verifica se ocorreu algum erro
-   if(!validationErros.isEmpty()){
-      fs.unlink(imgPath, (error) => {
-         if(error){
-            res.status(404).json(error);
-         }
-      });
-      return res.status(400).json({ errors: validationErros.array() });
+   }else{
+
+      const { path: imgPath } = req.file;
+   
+      // Guarda os erros de validação
+      const validationErros = validationResult(req);
+   
+      // Verifica se ocorreu algum erro
+      if(!validationErros.isEmpty()){
+         fs.unlink(imgPath, (error) => {
+            if(error){
+               res.status(404).json(error);
+            }
+         });
+         return res.status(400).json({ errors: validationErros.array() });
+      }
+
    }
 
    await Vehicles.addVehicle(req, res);
+   
 
 });
 
 router.patch("/vehicles/:id", login, upload.single("img"), validationRulesEdit, async (req, res) => {
 
-   const { path: imgPath } = req.file;
+   // Verifica se foi passada uma img via form
+   if(!req.file){
 
-   // Guarda os erros de validação
-   const validationErros = validationResult(req);
+      // Guarda os erros de validação
+      const validationErros = validationResult(req);
+         
+      // Verifica se ocorreu algum erro
+      if(!validationErros.isEmpty()){
+         return res.status(400).json({ errors: validationErros.array() });
+      }
 
-   // Verifica se ocorreu algum erro
-   if(!validationErros.isEmpty()){
-      fs.unlink(imgPath, (error) => {
-         if(error){
-            res.status(404).json(error);
-         }
-      });
-      return res.status(400).json({ errors: validationErros.array() });
+   }else{
+
+      const { path: imgPath } = req.file;
+   
+      // Guarda os erros de validação
+      const validationErros = validationResult(req);
+   
+      // Verifica se ocorreu algum erro
+      if(!validationErros.isEmpty()){
+         fs.unlink(imgPath, (error) => {
+            if(error){
+               res.status(404).json(error);
+            }
+         });
+         return res.status(400).json({ errors: validationErros.array() });
+      }
+      
    }
 
    await Vehicles.editVehicle(req, res);
